requirements.txt
streamlit
playwright
gspread
python-dotenv
google-auth
pandas
beautifulsoup4
google-generativeai
fastapi
uvicorn
pytest-playwright
lxml
requests
